Rewrite Blender export script
